migrating to a context to avoid re-render loops when creating multiple instances of the hook
