+++ frontend/src/App.js
@@ -1,11 +1,16 @@
-import React from "react";
+import React, { useEffect } from "react";
 import Sky from "./components/sky";
 import Loader from "./components/loader";
 import Rockets from "./components/rockets-list";
-import { useRockets } from "./hooks/useRockets";
+import { RocketsProvider, useRocketsContext } from "./contexts/RocketsContext";
 
-function App() {
-  const { rockets, maxRocketDimensions, loading, error } = useRockets();
+function AppContent() {
+  const { rockets, maxRocketDimensions, loading, error, fetchRockets } =
+    useRocketsContext();
+
+  useEffect(() => {
+    fetchRockets();
+  }, []);
 
   return (
     <div className="App">
@@ -14,6 +19,14 @@ function App() {
       {error && <div>{error}</div>}
       {maxRocketDimensions !== undefined && <Rockets rockets={rockets} />}
     </div>
+  );
+}
+
+function App() {
+  return (
+    <RocketsProvider>
+      <AppContent />
+    </RocketsProvider>
   );
 }
